bridging {influenceR}	R Documentation

Valente's Bridging vertex measure.

Description

Edges that reduces distances in a network are important structural bridges. Here we implement Valente and Fujimoto's metric, where a node's bridging score is the average decrease in cohesiveness if each of its edges were removed from the graph.

Usage

bridging(g)

Arguments

g
The igraph object to analyze.

Value

A numeric vector with the bridging score for each vertex

References

https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2889704/

Examples

ig.ex <- igraph::erdos.renyi.game(100, p.or.m=0.3) # generate an undirected 'igraph' object
bridging(ig.ex) # bridging scores for each node in the graph

[Package influenceR version 0.1.0.1 Index]