ens {influenceR}	R Documentation

Burt's Effective Network Size and Constraint index. The next two functions below provide ways to measure the actors' access to structural holes in a network. Structural holes "provide opportunities to broker connections between people" (Burt 2008).

Description

Burt's Effective Network Size and Constraint index. The next two functions below provide ways to measure the actors' access to structural holes in a network. Structural holes "provide opportunities to broker connections between people" (Burt 2008).

Usage

ens(g)

Arguments

g
The igraph object to analyze.

Value

A numeric vector with the effective network size for each vertex

References

http://ronaldsburt.com/research/files/NNappB.pdf

Examples

ig.ex <- igraph::erdos.renyi.game(100, p.or.m=0.3) # generate an undirected 'igraph' object
ens(ig.ex) # Effective Network Size scores for each node in the graph

[Package influenceR version 0.1.0.1 Index]